Safari-609.1.11.1/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/fetch/security/embedded-credentials.tentative.sub-expected.txt

Blocked access to external URL http://user:pass@www.localhost:8800/images/red.png
Blocked access to external URL http://user:pass@www.localhost:8800/images/red.png
Blocked access to external URL http://user:pass@www.localhost:8800/images/red.png

PASS Embedded credentials are treated as network errors. 
PASS Embedded credentials are treated as network errors in frames. 
PASS Embedded credentials are treated as network errors in new windows. 
PASS Embedded credentials matching the top-level are not treated as network errors for relative URLs. 
PASS Embedded credentials matching the top-level are not treated as network errors for same-origin URLs. 
FAIL Embedded credentials matching the top-level are treated as network errors for cross-origin URLs. assert_equals: The image should load. expected "Error" but got "Load"
